Remove complex variants feature
The variants feature wasn’t really useful and unnecessary complex. We are removing the feature in favor of a simple Snipcart API call to get the variants. So you could still show a variants stock in your frontend with some custom JS.

We also moved all custom field options into a single replicator for simplicity. This also makes it possible for you to control the order of the fields in the cart.

We also generally optimized the blueprint and added an inventory tab that shows the stock and variants.

The custom fields are no longer localizable too. This was super complicated and never lead to a desired result. We will bring this feature back, when Snipcart adds localizable custom fields and price modifiers per currency.